Conflict in Reducing Database Inventories
Reducing database inventories could incentivize the wrong behavior — In
existing environments agencies may have consolidated databases to a
single monolithic server - which can only scale up to meet demand and
is expensive and risky to maintain.

In a modern 12-factor app (http://12factor.net/) we would actually
decouple a centralize database dependency so in a virtual environment
database counts will go _up_ not down; which is actually a good thing.